Extract chat summaries
